clean lint

+++ src/interfaces/IynEigen.sol
@@ -2,9 +2,6 @@
 pragma solidity ^0.8.24;
 
 import {IERC20} from  "lib/openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
-import {IStrategyManager,IStrategy} from "lib/eigenlayer-contracts/src/contracts/interfaces/IStrategyManager.sol";
-import {IDelegationManager} from "lib/eigenlayer-contracts/src/contracts/interfaces/IDelegationManager.sol";
-import {UpgradeableBeacon} from "lib/openzeppelin-contracts/contracts/proxy/beacon/UpgradeableBeacon.sol";
 
 interface IynEigen {
 
@@ -14,9 +11,9 @@ interface IynEigen {
         address receiver
     ) external returns (uint256 shares);
 
-    function totalAssets() external view returns (uint);
+    function totalAssets() external view returns (uint256);
 
-    function convertToShares(IERC20 asset, uint amount) external view returns(uint shares);
+    function convertToShares(IERC20 asset, uint256 amount) external view returns(uint256 shares);
 
     function retrieveAssets(
         IERC20[] calldata assetsToRetrieve,
